feat(pwa): update players and pairs management
